一旦コンフリクトを修正してある程度ファイルとメソッドを整理
